refactor: removes irrelevant changes
